+++ app/src/main/java/com/example/thumbup/Fragment_home.java
@@ -1,0 +1,11 @@
+package com.example.thumbup;
+
+import android.view.View;
+
+import androidx.fragment.app.Fragment;
+
+public class Fragment_home extends Fragment {
+    private View view;
+
+
+}
