+++ app/src/main/res/drawable/ic_home_background.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="375dp"
+    android:height="757dp"
+    android:viewportWidth="375"
+    android:viewportHeight="757">
+  <path
+      android:pathData="M106,0H0V757H375V50H155.5C147.167,49.667 133.948,49.701 130.5,30C127,0 106,0 106,0Z"
+      android:fillColor="#D6E1F5"/>
+</vector>
