+++ app/src/main/res/drawable/ic_home_meeting_list_menu.xml
@@ -1,0 +1,27 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="25dp"
+    android:viewportWidth="24"
+    android:viewportHeight="25">
+  <path
+      android:pathData="M2,12.5857H22"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.8"
+      android:fillColor="#00000000"
+      android:strokeColor="#262F56"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M2,18.5857H22"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.8"
+      android:fillColor="#00000000"
+      android:strokeColor="#262F56"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M2,6.5857H22"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.8"
+      android:fillColor="#00000000"
+      android:strokeColor="#262F56"
+      android:strokeLineCap="round"/>
+</vector>
