+++ app/src/main/res/values-v21/themes.xml
@@ -1,7 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:tools="http://schemas.android.com/tools">
-    <!-- Base application theme. -->
-    <style name="Theme.ThumbUp" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
 
+    <style name="Theme.ThumbUp" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+        <item name="android:windowTranslucentStatus">true</item>
+        <item name="android:windowTranslucentNavigation">true</item>
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/main</item>
         <item name="colorOnPrimary">@color/white</item>
