+++ app/src/main/res/values/colors.xml
@@ -5,6 +5,11 @@
     <color name="purple_700">#FF3700B3</color>
     <color name="teal_200">#FF03DAC5</color>
     <color name="teal_700">#FF018786</color>
-    <color name="black">#FF000000</color>
+
+    <color name="main">#D6E1F5</color>
+    <color name="navy">#303F80</color>
+    <color name="light_gray">#EDEDED</color>
+    <color name="gray">#D2D2D2</color>
+    <color name="black">#000000</color>
     <color name="white">#FFFFFFFF</color>
 </resources>
