+++ app/src/main/res/values/strings.xml
@@ -2,4 +2,10 @@
     <string name="app_name">ThumbUp</string>
     <string name="facebook_app_id">785235652097318</string>
     <string name="fb_login_protocol_scheme">fb785235652097318</string>
+
+<!--  home  -->
+    <!--  margin  -->
+    <string name="home_marginRight_meeting_list_menu">12dp</string>
+    <!--  padding  -->
+    <string name="home_padding_LinearLayout">16dp</string>
 </resources>
